requirements.txt
torch
torchvision
transformers
peft
accelerate
ultralytics
mediapipe
Pillow
numpy
tqdm
pyyaml
